+++ contracts/interfaces/IBondPositionManager.sol
@@ -35,7 +35,7 @@ interface IBondPositionManager is IERC721Metadata, IERC721Permit, IBondPositionM
     function isAssetRegistered(address asset) external view returns (bool);
 
     // User Functions //
-    function mint(address asset, uint256 amount, uint16 lockEpochs, address recipient, bytes calldata data)
+    function mint(address asset, uint256 amount, uint8 lockEpochs, address recipient, bytes calldata data)
         external
         returns (uint256);
